requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
driver==0.1.0
duckduckgo_search==8.0.3
emoji==2.14.1
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.115.12
filetype==1.2.0
fonttools==4.58.1
fpdf2==2.8.3
frozenlist==1.6.0
fuzzywuzzy==0.18.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflate64==1.0.1
joblib==1.5.1
kaitaistruct==0.10
langdetect==1.0.9
Levenshtein==0.27.1
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.4
multivolumefile==0.2.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
olefile==0.47
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pi_heif==0.22.0
pillow==11.2.1
playwright==1.52.0
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py7zr==0.22.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
PyMuPDF==1.26.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyppmd==1.1.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-Levenshtein==0.27.1
python-magic==0.4.27
python-oxmsg==0.0.2
python-slugify==8.0.4
python-telegram-bot==22.1
pytz==2025.2
pyzstd==0.17.0
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
selenium==4.33.0
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.46.2
text-unidecode==1.3
texttable==1.7.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
